cnf-tests: introduce hugepages-allocator
Some of the tests running under cnf-tests are using the
`libhugetlbfs` for allocating hugepages.

`libhugetlbfs` has been removed from rhel9,
so instead we introduce a small program named
hugepages-allocator that will be used by the tests
for hugepages allocations.

The program built like any other tool we already
have in cnf-tests and the binary delivered as
part of the final image.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>