cnf-tests: delete Dockerfile
This Dockerfile is using centos7 but all the binaries are built using rhel8
which is not compatible at all.

IOW building cnf-tests using this Dockerfile, result
at broken image.

OTOH, we cannot switch to rhel8 here because:
1. rhel8 is not avilable for free as centos7 is.
2. switching to rhel8 result with the exact same `Dockerfile` as `Dockerfile.openshift`.

Due to the above it is suggested to remove the Dockerfile completely
and spare the redundant maintenance effort.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>