docs: elaborated instructions for branching in readme.md
We've recently faced issues due to misaligned cnf-tests container images.
Since it's easy to forget updating the image version across the repo, these changes to the branching guide are necessary to cover (hopefully) all required updates when creating a new release branch.
The intention is for anyone creating the release-x.y branch to follow this, minimizing the chances of misalignments.
Additionally, we implemented automation in the CI that will trigger a failure if these instructions are not followed: https://github.com/openshift-kni/cnf-features-deploy/pull/2199.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>