cnf-tests: create CI pipelines
Add yaml files to configure the integration tests of Konflux pointing to
the dockerfile in cnf-tests/.konflux.
Some binaries that are part of the build process needed to be prefetched,
prefetching dependencies is done via a flow called hermetic build in which
packages are pulled in a network-isolated way based on already generated
signatures stored in `rpms.lock.yaml` file. see
https://github.com/openshift-kni/cnf-features-deploy/pull/2167#issuecomment-2658977247
for additional details.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>