cnf-tests:k5x: add missing `varsFromContainerfile`
In order to interpolate {version} in the input file, pass
`varsFromContainerfile: Dockerfile` under each repo item. This will
follow the container image parameters that is used as the final image.
Without this parameter the re-generation of teh lockfile will fail,
eventhough the intergration test didn't catch this. Konflux is looking
at improving this validation but still fix this here.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>