infra: renovate: update submodules by default schedule
we've been seeing that updates of tekton and submodules are combined in
one PR. tekton file updates of the tasks digest do not require reviews,
so safe to automerge those. while the submodule updates are different as
they are used to run the prow ci.
such updates can be done per the default renovate schedule (ref:
https://konflux.pages.redhat.com/docs/users/mintmaker/user.html#scheduling
once a week) instead of at any indicated change on the modules repo.
This will alow another layer of difference between tekton and
submodules, which help decrease the amount of PRs for such updates.

Note that submodule updates are disabled by default, thus enable them
explicitly.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>