cnf-tests: k5x: bump rhel 9.4 -> 9.6
The health index for rhel9 images that are not based on the latest rhel9
minor versions are very low (D). Compared with the old method that used to
build the container images there was an automatic mechanism enabling it
to dynamically bump the base images such that the base will always be
the most recent. It was noticed that using the old method, the builds
used 9.6 and are A-scored. This commit bumps the ubi to 9.6
Considering there is no risk in that because cnf-tests is shipping tests
only.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>