ztp: Maintain legacy symlink layout in the ztp-site-generate Containerfile
The script executed as part of the container build takes the locked-down
manifest recorded in the `source-crs_layout_4.19.txt` file, and ensures
that if those files do not exist in source-crs exactly as named, we
manufacture symlinks pointing into whatever deep hierarchy source-crs
has changed into.

It will fail if there are missing files, or if there are duplicately
named files in the new hierarchy (though it does handle the special case
where if there are duplicately-named files and one contains 'x86' we
assume there's an architecture-specific distinction and pick the one
with 'x86' in the name).

Signed-off-by: Jim Ramsay <jramsay@redhat.com>