cnf-tests: OCPBUGS-58567: CVE-2024-45339: remove glog usage in application code
remove glog as direct dependency. The couple places
on which we called it, we can use other more common dependencies
to achieve the same result.

We still need glog as indirect dependency, so we
bump the required version to v1.2.5 to fix CVE per summary.

Is unlikely that vulnerability can ever be exploited in this code,
especially because we don't ship it externally anymore.

Signed-off-by: Francesco Romani <fromani@redhat.com>